DownwardMetrics: Add virtio-serial server
Adds the virtio-serial DownwardMetrics server, it provides a "metrics
communication channel" using a virtio-serial device to transport
the metrics data between the host and the VM.

The socket is created by QEMU, so the DownwardMetrics server must
connect to it. The connection will fail after few connection attempts,
waiting some time between attempts.

We also considered other options, have the server create the socket and
QEMU connect to it, and use a named pipe instead of a socket.
The former has the problem that we need to create the socket after the
launcher runs but before QEMU runs, or QEMU may fail due to timeout.
The latter, does not have that problem, but it changes the behavior of
the character device when there is nobody listening on the other side.
If a request is sent and no one is listening, the app client, in the
guest, will read back its own request.

Signed-off-by: Javier Cano Cano <jcanocan@redhat.com>
Signed-off-by: German Maglione <gmaglione@redhat.com>