DownwardMetrics: Add DownwardMetrics servers manager
DownwardMetricsManager is in charge of starting and stopping each
DownwardMetrics servers. The lifetime of each server is tied to the VMI
and the manager itself. The DownwardMetricsManager will watch for VMI
Update and Delete events in its node, and will launch and stop a
metrics server for each of VMI.

Signed-off-by: Javier Cano Cano <jcanocan@redhat.com>
Signed-off-by: German Maglione <gmaglione@redhat.com>